Whatsapp-empleo_camila/requirements.txt

playwright==1.47.0
python-dotenv==1.0.1
GitPython==3.1.43
pandas==2.2.2
beautifulsoup4==4.12.3
